+++ packages/js/api-core-tests/data/index.js
@@ -1,7 +1,9 @@
-const { order, getOrderExample } = require('./order');
-const { coupon } = require('./coupon');
-const { refund } = require('./refund');
-const shared = require('./shared');
+const { order, getOrderExample } = require( './order' );
+const { coupon } = require( './coupon' );
+const { refund } = require( './refund' );
+const { getShippingZoneExample } = require( './shipping-zone' );
+const { getShippingMethodExample } = require( './shipping-method' );
+const shared = require( './shared' );
 
 module.exports = {
 	order,
@@ -9,4 +11,6 @@ module.exports = {
 	coupon,
 	shared,
 	refund,
+	getShippingZoneExample,
+	getShippingMethodExample,
 };