+++ packages/js/api-core-tests/CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+## Added
+- Shipping Zones API Tests
+- Shipping Methods API Tests
+
 # 0.1.0
 
 - Initial/beta release
@@ -7,3 +13,4 @@
 - Refunds API Tests
 - Products API Tests
 - CRUD tests for the Orders API
+- Order Search API Tests